refactor: 패키지 구조 개선 및 패키지 이름 변경
