+++ src/main/java/racinggame/Application.java
@@ -1,8 +1,8 @@
-package game;
+package racinggame;
 
-import game.ui.RacingConsole;
-import game.ui.RacingInputImpl;
-import game.ui.RacingOutputImpl;
+import racinggame.ui.RacingConsole;
+import racinggame.ui.RacingInputImpl;
+import racinggame.ui.RacingOutputImpl;
 
 public class Application {
     public static void main(String[] args) {
