+++ src/main/java/racinggame/ui/RacingOutputImpl.java
@@ -1,6 +1,6 @@
-package game.ui;
+package racinggame.ui;
 
-import game.domain.GameResult;
+import racinggame.domain.dto.GameResult;
 
 public class RacingOutputImpl implements RacingOutput {
     @Override
